requirements.txt
pandas>=1.5.0
numpy>=1.21.0
ta>=0.10.0
python-binance>=1.0.0
python-dotenv>=0.19.0
scikit-learn>=1.0.2
joblib>=1.1.0
matplotlib>=3.5.0 
apscheduler>=3.9.1